Many Small changes to spells and animations.
Added Meteor Spell, Fixed Spell UI, Added Soldier Commander Attack Animation, added spell struct.

+++ RogueLikeGame/Assets/Scripts/HealthCheck.cs
@@ -8,7 +8,7 @@ public class HealthCheck : MonoBehaviour
 {
     public void updateHealth(PlayerClass pc)
     {
-        transform.Find("HealthText").GetComponent<Text>().text = pc.curHP.ToString();
+        transform.Find("HealthText").GetComponent<Text>().text = ((int)pc.curHP).ToString();
         transform.Find("HealthBar").GetComponent<RectTransform>().SetSizeWithCurrentAnchors(RectTransform.Axis.Horizontal, (Convert.ToSingle(pc.curHP) / pc.maxHP) * 200);
         
     }
